Post release automated changes for azure-communication-network-traversal
