+++ sdk/communication/communication-network-traversal/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@azure/communication-network-traversal",
-  "version": "1.0.0-beta.3",
+  "version": "1.0.0-beta.4",
   "description": "SDK for Azure Communication service which facilitates user token administration.",
   "sdk-type": "client",
   "main": "dist/index.js",
