+++ sdk/communication/communication-network-traversal/samples/v1/javascript/getRelayConfigurationWithIdentity.js
@@ -2,27 +2,28 @@
 // Licensed under the MIT License.
 
 /**
- * @summary Issue a new Relay configuration
+ * @summary Issue a new Relay configuration providing a user identity
  */
 
-import { CommunicationIdentityClient } from "@azure/communication-identity";
-import { CommunicationRelayClient } from "@azure/communication-network-traversal";
+const { CommunicationIdentityClient } = require("@azure/communication-identity");
+const { CommunicationRelayClient } = require("@azure/communication-network-traversal");
 
 // Load the .env file if it exists
-import * as dotenv from "dotenv";
+const dotenv = require("dotenv");
 dotenv.config();
 
 // You will need to set this environment variables or edit the following values
 const connectionString =
   process.env["COMMUNICATION_CONNECTION_STRING"] || "<communication service connection string>";
 
-export async function main() {
+async function main() {
   console.log("\n== Get Relay configuration Sample ==\n");
 
   const identityClient = new CommunicationIdentityClient(connectionString);
 
   // Create user
   console.log("Creating User");
+
   const user = await identityClient.createUser();
 
   const relayClient = new CommunicationRelayClient(connectionString);
