+++ sdk/communication/communication-network-traversal/samples/v1/javascript/getRelayConfigurationWithoutIdentity.js
@@ -2,10 +2,9 @@
 // Licensed under the MIT License.
 
 /**
- * @summary Issue a new Relay configuration
+ * @summary Issue a new Relay configuration without user identity
  */
 
-const { CommunicationIdentityClient } = require("@azure/communication-identity");
 const { CommunicationRelayClient } = require("@azure/communication-network-traversal");
 
 // Load the .env file if it exists
@@ -17,19 +16,12 @@ const connectionString =
   process.env["COMMUNICATION_CONNECTION_STRING"] || "<communication service connection string>";
 
 async function main() {
-  console.log("\n== Get Relay configuration Sample ==\n");
-
-  const identityClient = new CommunicationIdentityClient(connectionString);
-
-  // Create user
-  console.log("Creating User");
-
-  const user = await identityClient.createUser();
+  console.log("\n== Get Relay configuration without identity sample ==\n");
 
   const relayClient = new CommunicationRelayClient(connectionString);
   console.log("Getting relay configuration");
 
-  const config = await relayClient.getRelayConfiguration(user);
+  const config = await relayClient.getRelayConfiguration();
   console.log("RelayConfig", config);
 }
