+++ sdk/communication/communication-network-traversal/samples/v1/javascript/package.json
@@ -3,7 +3,7 @@
   "private": true,
   "version": "1.0.0",
   "description": "Azure Communication Services - Network Traversal client library samples for JavaScript",
-  "engine": {
+  "engines": {
     "node": ">=12.0.0"
   },
   "repository": {
