+++ sdk/communication/communication-network-traversal/samples/v1/typescript/package.json
@@ -3,7 +3,7 @@
   "private": true,
   "version": "1.0.0",
   "description": "Azure Communication Services - Network Traversal client library samples for TypeScript",
-  "engine": {
+  "engines": {
     "node": ">=12.0.0"
   },
   "scripts": {
@@ -32,7 +32,7 @@
     "@azure/communication-identity": "^1.0.0"
   },
   "devDependencies": {
-    "typescript": "~4.2.0",
+    "typescript": "~4.4.0",
     "rimraf": "latest"
   }
 }
